python/requirements.txt
fastapi==0.128.3
uvicorn==0.40.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
alembic==1.18.3
strawberry-graphql==0.291.2
pydantic==2.12.5
python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.22
email-validator==2.3.0
python-dotenv==1.2.1
strawberry-sqlalchemy-mapper==0.8.0
boto3==1.42.44
sendgrid==6.12.5
squareup==44.0.0.20260122
cryptography==46.0.4
apscheduler==3.11.2
cognitojwt==1.4.1
slowapi==0.1.9
pyjwt[crypto]==2.11.0
requests==2.32.5
